requirements.txt
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.9.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.4.0
astroid==2.11.7
asttokens==2.2.1
async-timeout==4.0.2
attrs==20.2.0
backcall==0.2.0
backoff==2.1.2
bandit==1.7.4
bcrypt==3.1.7
beautifulsoup4==4.11.1
bingads==13.0.11.1
black==22.1.0
bleach==3.3.0
boto3==1.26.105
botocore==1.29.105
bytecode==0.13.0
cached-property==1.5.2
cachetools==4.1.1
cchardet==2.1.7
certifi==2022.12.7
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
cmd2==1.0.2
colorama==0.4.3
colorlog==6.6.0
commonmark==0.9.1
coreapi==2.0.0
coverage==6.4.4
cryptography==39.0.1
datadog==0.44.0
ddsketch==2.0.3
ddtrace==1.2.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.5.1
distlib==0.3.1
Django==3.2.18
django-configurations==2.4
django-cors-headers==3.13.0
django-extensions==3.2.0
django-formtools==2.4
django-log-request-id==2.0.0
django-redis==5.2.0
django-rq==2.6.0
django-ses==3.1.0
django-slowtests==1.1.1
django-storages==1.13.2
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
expiringdict==1.1.4
factory-boy==3.2.1
Faker==2.0.0
filelock==3.5.0
flake8==3.9.2
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.30
gnureadline==8.0.0
google-ads==19.0.0
google-api-core==2.10.1
google-api-python-client==1.12.10
google-auth==1.25.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.3.0
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleads==20.0.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.46.0
gunicorn==20.1.0
hashin==0.17.0
honcho==1.1.0
httplib2==0.19.0
identify==1.5.9
idna==2.8
importlib-metadata==4.11.2
importlib-resources==1.4.0
inflection==0.3.1
iniconfig==2.0.0
install==1.3.5
invoke==1.7.1
ipykernel==5.1.1
ipython==8.10.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.12.0
itypes==1.2.0
jedi==0.17.2
Jinja2==3.0.1
jinjasql==0.1.8
jmespath==0.10.0
jsonschema==4.17.3
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.2.2
launchdarkly-server-sdk==7.5.0
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==0.8.4
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.6.7
nbconvert==6.5.1
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
nodeenv==1.5.0
notebook==6.4.10
nox==2022.1.7
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.0.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandas-stubs==1.4.4.220912
pandocfilters==1.4.3
parso==0.7.1
path.py==12.0.1
pathspec==0.9.0
pbr==5.4.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-api==0.0.26
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.17.0
pre-commit-hooks==4.3.0
prettytable==3.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.38
proto-plus==1.22.1
protobuf==4.21.6
psycopg2==2.9.3
ptyprocess==0.6.0
pure-eval==0.2.2
pusher==3.2.0
py==1.11.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.6
pycodestyle==2.7.0
pycparser==2.19
pyflakes==2.3.1
pygame==2.3.0
Pygments==2.7.4
PyJWT==2.4.0
pylint==2.14.5
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyperclip==1.8.1
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.0
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.5
python-json-logger==2.0.4
python3-openid==3.1.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==19.0.2
ratelimit==2.2.1
redis==4.5.4
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
requests-futures==1.0.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rich==12.4.4
rq==1.10.1
rsa==4.7
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.14.0
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.6.2
stevedore==3.5.0
suds-community==1.0.0
tblib==1.7.0
tenacity==8.0.1
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.2
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
tomlkit==0.11.2
tornado==6.1
tqdm==4.32.2
traitlets==5.3.0
tweepy==3.7.0
typed-ast==1.5.1
types-Pillow==9.4.0.6
types-pytz==2022.2.1.0
types-PyYAML==6.0.4
types-requests==2.27.7
types-urllib3==1.26.7
typing_extensions==4.0.1
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.12
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.0.3
xmltodict==0.12.0
zeep==3.4.0
zipp==3.4.0
